requirements.txt
ccxt==4.2.85
websockets==12.0
python-dotenv==1.0.0
python-telegram-bot==20.1
